Fix hiding of join/part/quit messages.
As it turns out, if you hide your own joins, parts, and quits, Riece gets
really confused about what channels you're on and wouldn't update the
channel list.  This changeset fixes that by not hiding the user's joins,
parts, or quits.

The variable controlling it was also renamed to something nicer.  ERC uses
`erc-hide-list' for this, so `riece-hide-list' seems a good fit.

	* lisp/riece-options.el (riece-hide-list): Renamed from
	`riece-hide-joins-parts-quits' which was way too ugly.

	* riece-handle.el (riece-handle-join-message): Use renamed hide
	var, don't hide our own joins.
	(riece-handle-part-message): Use renamed hide var, don't hide our
	own parts.
	(riece-handle-quit-message): Use renamed hide var, don't hide our
	own quits.

Signed-off-by: Steve Youngs <steve@sxemacs.org>